git-svn-id: http://192.168.20.11/sequoiadb/trunk@13911 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/omagent/removeVirtualCoord.js
@@ -1,14 +1,17 @@
-// installVirtualCoord
+/* *****************************************************************************
+@description: remove virtual coord
+@modify list:
+   2014-7-26 Zhaobo Tan  Init
+***************************************************************************** */
 if ( typeof(IP) == "undefined" ) { IP = "127.0.0.1" ; }
-if ( typeof(USERNAME) == "undefined" ) { USERNAME = "sdbadmin" ; }
-if ( typeof(PASSWORD) == "undefined" ) { PASSWORD = "adbadmin" ; }
-//if ( typeof(TIMES) == "undefined" ) { TIMES = 3 ; }
+if ( typeof(USERNAME) == "undefined" ) {}
+if ( typeof(PASSWORD) == "undefined" ) {}
 // todo::add path in windows
 if ( typeof(PROGRAM) == "undefined" ) { PROGRAM_PATH = "/opt/sequoiadb/bin/sdbstop" ; }
 //if ( typeof(CONFIG_PATH) == "undefined" ) { CONFIG_PATH = "/tmp/virtualCoord" ; }
 //if ( typeof(CONFIG_FILE) == "undefined" ) { CONFIG_FILE = "/tmp/virtualCoord/sdb.conf" ; }
 if ( typeof(DB_PATH) == "undefined" ) { CONFIG = "/tmp/virtualCoord" ; }
-if ( typeof(LOCAL_SERVICE) == "undefined" ) { LOCAL_SERVICE = "11810" ; }
+if ( typeof(COORD_SERVICE) == "undefined" ) { COORD_SERVICE = "11810" ; }
 var objRet = new Object() ;
 
 objRet.Rc = 0 ;
@@ -18,10 +21,19 @@ function main()
 {
    try
    {
+      // check arguments
+      if ( typeof(USERNAME) == "undefined" ||
+          typeof(PASSWORD) == "undefined" )
+      {
+         objRet.Rc = -6 ;
+         objRet.Detail = "Invalid arguments" ;
+         return objRet ;
+      }
+
       // ssh to local
       var ssh = new Ssh( IP, USERNAME, PASSWORD ) ;
       // stop virtual coord
-      var cmd = PROGRAM + " -p " + LOCAL_SERVICE ;
+      var cmd = PROGRAM + " -p " + COORD_SERVICE ;
       ssh.exec( cmd ) ;
       // remove rubbish files
       File.remove( CONFIG_PATH ) ;
